% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Convert a DNA sequence to RNA}
\usage{
dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{A character string representing a DNA sequence}
}
\value{
A character string representing the RNA sequence with all "T" replaced by "U".
}
\description{
Convert a DNA sequence to RNA
}
\examples{
dna_to_rna("ATTGCT")
dna_to_rna("TACG")
}
