% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna.R
\name{generate_dna}
\alias{generate_dna}
\title{Generate a random DNA sequence}
\usage{
generate_dna(length)
}
\arguments{
\item{length}{Numeric. The length of the DNA sequence to generate.}
}
\value{
A character string representing the generated DNA sequence.
}
\description{
This function generates a random DNA sequence of a specified length,
composed of the nucleotides A, T, G, and C.
}
\examples{
generate_dna(10)
generate_dna(20)

}
