% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amino_acid_counts.R
\name{plot_amino_acid_counts}
\alias{plot_amino_acid_counts}
\title{Plot amino acid frequency in a sequence}
\usage{
plot_amino_acid_counts(sequence)
}
\arguments{
\item{sequence}{Character string, Amino acid sequence}
}
\value{
A ggplot object showing the frequency of each amino acid.
}
\description{
Plot amino acid frequency in a sequence
}
\examples{
plot_amino_acid_counts("MAFTIG")

}
