% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_into_codons.R
\name{split_into_codons}
\alias{split_into_codons}
\title{Split a nucleotide sequence into columns}
\usage{
split_into_codons(nucleotide_sequence, start = 1)
}
\arguments{
\item{nucleotide_sequence}{a sequence of nucleotides as string}

\item{start}{an integer number giving the start of the sequence}
}
\value{
codons a character vector containing codons
}
\description{
Split a nucleotide sequence into columns
}
\examples{
split_into_codons("ATGCGTACG")
#> [1] "ATG" "CGT" "ACG"
split_into_codons("ATGCGTACG", start = 2)
#> [1] "TGC" "GTA"

}
