% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_codons.R
\name{translate_codons}
\alias{translate_codons}
\title{Translate codons into an amino acid sequence}
\usage{
translate_codons(codons)
}
\arguments{
\item{codons}{A character vector containing codon triplets (e.g., c("AUG", "UUU", "GGC")).}
}
\value{
A single character string representing the translated amino acid sequence.
}
\description{
This function takes a vector of codons (e.g., "AUG", "UUU", "GGC")
and translates it into the corresponding amino acid sequence
using the `codon_table` stored in your package.
}
\examples{
codon_table <- c(AUG = "M", UUU = "F", GGC = "G")
translate_codons(c("AUG", "UUU", "GGC"))
# Expected output: "MFG"

}
